refactor code interface
